build 20200419 + refactor data bus to full width (8 bit -> 32 bit) + 1801VM1 disasm: E(R7) is now showing as (#addr) instead of #addr + Fix char attributes in ATM2 text mode

+++ src/xcore/common.cpp
@@ -21,7 +21,7 @@ QString gethexword(int num) {
 	return res.rightJustified(4, '0');
 }
 
-QString gethexbyte(uchar num) {
+QString gethexbyte(int num) {
 	QString res = formbufword(num & 0xff);
 	return res.rightJustified(2, '0');
 }
